# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Magenta ASOAR API
  description: |
    Agentic System Orchestration Automation and Response — REST API.

    Provides endpoints for agent management, mission orchestration,
    log ingestion, data mesh queries, memory operations, and system health.
  version: 0.1.0
  contact:
    name: Magenta SOC Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://magenta-api.magenta-soa.svc.cluster.local
    description: Kubernetes cluster
...
paths:
...
  /api/v1/playbooks:
    get:
      tags: [Playbooks]
      summary: List playbooks
      operationId: listPlaybooks
      parameters:
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Playbook list

    post:
      tags: [Playbooks]
      summary: Register a playbook
      operationId: registerPlaybook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Playbook registered
  /api/v1/playbooks/{name}:
    get:
      tags: [Playbooks]
      summary: Get playbook details
      operationId: getPlaybook
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Playbook details
  /api/v1/playbooks/validate:
    post:
      tags: [Playbooks]
      summary: Validate a playbook
      operationId: validatePlaybook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Validation result
...
components:
  securitySchemes:
    hmacAuth:
      type: apiKey
      in: header
      name: X-Magenta-Signature
      description: HMAC-SHA256 signature
...